+++ CHANGELOG.md
@@ -8,6 +8,52 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [5.31.7] - 2021-10-04
+
+### Changed
+
+- Bumped `@jupiterone/integration-sdk-*@6.22.1`. This included new functionality
+  to the `IntegrationError` classes that better segregates errors coming from
+  different steps
+
+## [5.31.6] - 2021-10-01
+
+### Fixed
+
+- Stop throwing
+  `Provider API failed at storage.*: AccountIsDisabled The specified account is disabled.`
+  on storage blob/queue/table/file steps
+
+## [5.31.5] - 2021-10-01
+
+### Fixed
+
+- Stop throwing
+  `Provider API failed at storage.*: AccountIsDisabled The specified account is disabled.`
+  on storage blob/queue/table/file steps
+
+## [5.31.5] - 2021-10-01
+
+### Fixed
+
+- Stop throwing
+  `Provider API failed at storage.*: AccountIsDisabled The specified account is disabled.`
+  on storage blob/queue/table/file steps
+- Stop throwing
+  `Provider API failed at *.*: DisallowedOperation The current subscription type is not permitted to perform operations on any provider namespace. Please use a different subscription.`
+  on steps
+
+## [5.31.4] - 2021-10-01
+
+### Fixed
+
+- Stop throwing
+  `Provider API failed at monitor.diagnosticSetting: SubscriptionNotRegistered The subscription '<SUBSCRIPTION-ID>' is not registered to use microsoft.insights.`
+  on security center settings step
+- Stop throwing
+  `Provider API failed at security.settings: Subscription Not Registered Please register to Microsoft.Security in order to view your security status`
+  on security center settings step
+
 ## [5.31.3] - 2021-09-07
 
 ### Changed